+++ stars-core/src/main/kotlin/tools/aqua/stars/core/metric/providers/Loggable.kt
@@ -20,10 +20,14 @@ package tools.aqua.stars.core.metric.providers
 import java.io.File
 import java.util.logging.*
 import tools.aqua.stars.core.metric.utils.ApplicationConstantsHolder
+import tools.aqua.stars.core.metric.utils.ApplicationConstantsHolder.activeLoggers
 
 /** This interface can be implemented to be able to log data into the stdout and log files. */
 @Suppress("unused")
 interface Loggable {
+  /** Holds the identifier (name) for this logger. */
+  val loggerIdentifier: String
+
   /** Holds the [Logger] reference for this class. */
   val logger: Logger
 
@@ -98,6 +102,7 @@ interface Loggable {
     logger.handlers.forEach { it.close() }
   }
 
+  /** Provides static function for the creation of [Logger]s. */
   companion object {
     /**
      * Creates a [Logger] with a [FileHandler] each for [Level.SEVERE], [Level.WARNING],
@@ -116,17 +121,19 @@ interface Loggable {
           }
       val file = "$logFolderFile/$name-${currentTimeAndDate}"
 
-      return@run Logger.getAnonymousLogger().apply {
-        useParentHandlers = false
-        level = Level.FINEST
+      return@run Logger.getAnonymousLogger()
+          .apply {
+            useParentHandlers = false
+            level = Level.FINEST
 
-        addHandler(getLoggerHandler(file, Level.SEVERE))
-        addHandler(getLoggerHandler(file, Level.WARNING))
-        addHandler(getLoggerHandler(file, Level.INFO))
-        addHandler(getLoggerHandler(file, Level.FINE))
-        addHandler(getLoggerHandler(file, Level.FINER))
-        addHandler(getLoggerHandler(file, Level.FINEST))
-      }
+            addHandler(getLoggerHandler(file, Level.SEVERE))
+            addHandler(getLoggerHandler(file, Level.WARNING))
+            addHandler(getLoggerHandler(file, Level.INFO))
+            addHandler(getLoggerHandler(file, Level.FINE))
+            addHandler(getLoggerHandler(file, Level.FINER))
+            addHandler(getLoggerHandler(file, Level.FINEST))
+          }
+          .also { activeLoggers.add(it) }
     }
 
     private fun getLoggerHandler(file: String, level: Level) =